+++ packages/categorize/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.2.2...@pie-ui/categorize@0.3.0) (2018-05-31)
+
+
+### Bug Fixes
+
+* use [@pie-lib](https://github.com/pie-lib)/drag@^1.0.0 ([e88b16a](https://github.com/pie-framework/pie-ui/commit/e88b16a))
+
+
+### Features
+
+* new component fill-in-the-blank ([d6129f6](https://github.com/pie-framework/pie-ui/commit/d6129f6))
+
+
+
+
+<a name="0.2.2"></a>
+## [0.2.2](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.2.1...@pie-ui/categorize@0.2.2) (2018-05-21)
+
+
+### Bug Fixes
+
+* use new [@pie-lib](https://github.com/pie-lib)/categorize package ([f7c479e](https://github.com/pie-framework/pie-ui/commit/f7c479e))
+
+
+
+
 <a name="0.2.1"></a>
 ## [0.2.1](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.2.0...@pie-ui/categorize@0.2.1) (2018-05-16)